changes in migration of service requests

+++ database/migrations/2023_11_11_112527_add_column_to_service_requests.php
@@ -14,7 +14,7 @@ return new class extends Migration
     public function up()
     {
         Schema::table('service_requests', function (Blueprint $table) {
-            $table->integer('req_number');
+            $table->bigInteger('req_number');
         });
     }
